requirements.txt
future==0.18.3
iso8601==2.1.0
numpy==1.26.3
pandas==2.2.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
RPi.GPIO==0.7.1
six==1.16.0
tzdata==2023.4
